chore: update example app to use @analogjs/router package
