--- apps/analog-app/src/app/cart/cart.component.css
@@ -1,2 +1,0 @@
-
-
