--- apps/analog-app/src/app/cart/cart.component.ts
@@ -1,34 +1,0 @@
-import { CommonModule } from '@angular/common';
-import { Component } from '@angular/core';
-import { FormBuilder, ReactiveFormsModule } from '@angular/forms';
-import { RouterModule } from '@angular/router';
-
-import { CartService } from '../cart.service';
-
-@Component({
-  selector: 'app-cart',
-  standalone: true,
-  imports: [CommonModule, ReactiveFormsModule, RouterModule],
-  templateUrl: './cart.component.html',
-  styleUrls: ['./cart.component.css'],
-})
-export class CartComponent {
-  items = this.cartService.getItems();
-
-  checkoutForm = this.formBuilder.group({
-    name: '',
-    address: '',
-  });
-
-  constructor(
-    private cartService: CartService,
-    private formBuilder: FormBuilder
-  ) {}
-
-  onSubmit(): void {
-    // Process checkout data here
-    this.items = this.cartService.clearCart();
-    console.warn('Your order has been submitted', this.checkoutForm.value);
-    this.checkoutForm.reset();
-  }
-}
