--- apps/analog-app/src/app/product-details/product-details.component.css
@@ -1,2 +1,0 @@
-
-
