--- apps/analog-app/src/app/product-list/product-list.component.html
@@ -1,28 +1,0 @@
-<h2>Products</h2>
-
-<div *ngFor="let product of products">
-
-  <h3>
-    <a
-      [title]="product.name + ' details'"
-      [routerLink]="['/products', product.id]">
-      {{ product.name }}
-    </a>
-  </h3>
-
-  <p *ngIf="product.description">
-    Description: {{ product.description }}
-  </p>
-
-  <button type="button" (click)="share()">
-    Share
-  </button>
-
-  <app-product-alerts
-    [product]="product"
-    (notify)="onNotify()">
-  </app-product-alerts>
-
-</div>
-
-
