--- apps/analog-app/src/app/routes.ts
@@ -1,26 +1,0 @@
-export const routes = [
-  {
-    path: '',
-    loadComponent: () =>
-      import('./product-list/product-list.component').then(
-        (m) => m.ProductListComponent
-      ),
-  },
-  {
-    path: 'products/:productId',
-    loadComponent: () =>
-      import('./product-details/product-details.component').then(
-        (m) => m.ProductDetailsComponent
-      ),
-  },
-  {
-    path: 'cart',
-    loadComponent: () =>
-      import('./cart/cart.component').then((m) => m.CartComponent),
-  },
-  {
-    path: 'shipping',
-    loadComponent: () =>
-      import('./shipping/shipping.component').then((m) => m.ShippingComponent),
-  },
-];
