--- apps/analog-app/src/app/shipping/shipping.component.css
@@ -1,2 +1,0 @@
-
-
