--- apps/analog-app/src/app/shipping/shipping.component.ts
@@ -1,26 +1,0 @@
-import { CommonModule } from '@angular/common';
-import { Component, OnInit } from '@angular/core';
-
-import { Observable } from 'rxjs';
-import { CartService } from '../cart.service';
-
-@Component({
-  selector: 'app-shipping',
-  standalone: true,
-  imports: [CommonModule],
-  templateUrl: './shipping.component.html',
-  styleUrls: ['./shipping.component.css']
-})
-export class ShippingComponent implements OnInit {
-
-  shippingCosts!: Observable<{ type: string, price: number }[]>;
-
-  constructor(private cartService: CartService) { }
-
-  ngOnInit(): void {
-    this.shippingCosts =  this.cartService.getShippingPrices();
-  }
-
-}
-
-
