+++ apps/analog-app/src/routes/product-details.component.html
