+++ apps/analog-app/src/routes/products.[productId].ts
@@ -1,0 +1,39 @@
+import { Component, OnInit } from '@angular/core';
+import { ActivatedRoute } from '@angular/router';
+import { CommonModule } from '@angular/common';
+
+import { Product, products } from '../app/products';
+import { CartService } from '../app/cart.service';
+
+@Component({
+  selector: 'app-product-details',
+  standalone: true,
+  imports: [CommonModule],
+  templateUrl: './product-details.component.html',
+  styleUrls: ['./product-details.component.css'],
+})
+export default class ProductDetailsComponent implements OnInit {
+  product: Product | undefined;
+
+  constructor(
+    private route: ActivatedRoute,
+    private cartService: CartService
+  ) {}
+
+  ngOnInit() {
+    // First get the product id from the current route.
+    console.log(this.route);
+    const routeParams = this.route.parent!.snapshot!.paramMap;
+    const productIdFromRoute = Number(routeParams.get('productId'));
+
+    // Find the product that correspond with the id provided in route.
+    this.product = products.find(
+      (product) => product.id === productIdFromRoute
+    );
+  }
+
+  addToCart(product: Product) {
+    this.cartService.addToCart(product);
+    window.alert('Your product has been added to the cart!');
+  }
+}
