+++ apps/analog-app/src/routes/shipping/index.ts
@@ -1,0 +1,22 @@
+import { CommonModule } from '@angular/common';
+import { Component, OnInit } from '@angular/core';
+
+import { Observable } from 'rxjs';
+import { CartService } from '../../app/cart.service';
+
+@Component({
+  selector: 'app-shipping',
+  standalone: true,
+  imports: [CommonModule],
+  templateUrl: './shipping.component.html',
+  styleUrls: ['./shipping.component.css'],
+})
+export default class ShippingComponent implements OnInit {
+  shippingCosts!: Observable<{ type: string; price: number }[]>;
+
+  constructor(private cartService: CartService) {}
+
+  ngOnInit(): void {
+    this.shippingCosts = this.cartService.getShippingPrices();
+  }
+}
